+++ tag-release.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-if [ $# -lt 2 ] ; then
+if [ $# -lt 1 ] ; then
     echo "usage : `basename $0` v1.0.0";
     exit -1
 fi
@@ -15,7 +15,7 @@ set -e
 carthage build --no-skip-current && carthage archive SwiftiumKit
 
 echo "Creating tag $TAG and pushing it to github"
-git tag -a $TAG
+git tag $TAG
 git push --tags
 
 echo "You can now upload SwiftiumKit.framework.zip and edit release notes from https://github.com/openium/SwiftiumKit/releases/edit/$TAG"